make different food truly valid and fix the bug (Copilot new feature! :) The most significant changes were made to the `ancillarycat.game.game.ixx` file, which underwent a major refactoring. The `score_logger` and `time_logger` functions were modified, the `gameOver` and `gameInit` functions were rewritten and moved, and the `snakeGame` function was completely rewritten. New functions `input_key` and `make_valid_food` were added, and new variables `snake`, `foods`, `isIgnore`, and `prevKey` were introduced. The `gameOver` function that took a `Snake` object as a parameter was removed, and the `checkInvalidPosition` function was modified to take const references as parameters. A bug was identified in the `instanceof` function.
Other important changes include updates to the `Entity` and `Food` classes in `ancillarycat.entities.entity.ixx` and `ancillarycat.entities.food.ixx` respectively, where the constructor and destructor were changed and a new function `count()` was added. The `generate_food` function in `ancillarycat.entities.ixx` was updated to return a `std::unique_ptr<Food>` instead of a `std::vector<Food>`. The `getch` function in both `ancillarycat.console.ixx` and `ancillarycat.console.cpp` was updated to change a parameter name. The `CMakeLists.txt` file was updated to set the C++ standard to 23.

In `ancillarycat.handler.cpp`, the `oninitialize` function was modified to remove a statement. In `ancillarycat.utils.cpp`, the `checkInvalidPosition` function was modified to take const references as parameters and to use methods instead of directly accessing properties. The checks for `instanceof<Entity, Snake>` and `instanceof<Entity, Food>` were commented out, and a bug was identified in the `instanceof` function.

List of changes:

1. `CMakeLists.txt`: Set C++ standard to 23.
2. `ancillarycat.console.ixx` and `ancillarycat.console.cpp`: Updated `getch` function parameter name.
3. `ancillarycat.entities.entity.ixx` and `ancillarycat.entities.food.ixx`: Updated `Entity` and `Food` classes.
4. `ancillarycat.entities.ixx`: Updated `generate_food` function return type.
5. `ancillarycat.game.ixx`: Updated `gameOver` function parameter type, and major refactoring of `ancillarycat.game.game.ixx`.
6. `ancillarycat.utils.ixx`: Updated `checkInvalidPosition` function parameter type.
7. `ancillarycat.handler.cpp`: Modified `oninitialize` function.
8. `ancillarycat.utils.cpp`: Modified `checkInvalidPosition` function and commented out checks for `instanceof<Entity, Snake>` and `instanceof<Entity, Food>`.